LPS-127552 Move openFormViewModal to EditApp context
