+++ modules/dxp/apps/app-builder-workflow/app-builder-workflow-web/src/main/resources/META-INF/resources/js/pages/apps/edit/sidebar/DataAndViewsTab.es.js
@@ -18,7 +18,6 @@ import {AppContext} from 'app-builder-web/js/AppContext.es';
 import SelectObjects from 'app-builder-web/js/components/select-objects/SelectObjects.es';
 import EditAppContext, {
 	UPDATE_APP,
-	UPDATE_DATA_LAYOUT_ID,
 	UPDATE_DATA_LIST_VIEW_ID,
 } from 'app-builder-web/js/pages/apps/edit/EditAppContext.es';
 import {getLocalizedValue, sub} from 'app-builder-web/js/utils/lang.es';
@@ -27,14 +26,13 @@ import {concatValues} from 'app-builder-web/js/utils/utils.es';
 import classNames from 'classnames';
 import {DataDefinitionUtils} from 'data-engine-taglib';
 import {openModal} from 'frontend-js-web';
-import React, {createContext, useContext} from 'react';
+import React, {useContext} from 'react';
 
-import {getFormViews, getTableViews} from '../actions.es';
+import {getTableViews} from '../actions.es';
 import {
 	ADD_STEP_FORM_VIEW,
 	REMOVE_STEP_FORM_VIEW,
 	UPDATE_DATA_OBJECT,
-	UPDATE_FORM_VIEW,
 	UPDATE_LIST_ITEMS,
 	UPDATE_STEP_FORM_VIEW,
 	UPDATE_STEP_FORM_VIEW_READONLY,
@@ -43,8 +41,6 @@ import {
 import {checkRequiredFields} from '../utils.es';
 import SelectFormView from './SelectFormView.es';
 import SelectTableView from './SelectTableView.es';
-
-export const DataAndViewsTabContext = createContext();
 
 const NoObjectEmptyState = () => (
 	<div className="taglib-empty-result-message">
@@ -91,7 +87,9 @@ export default function DataAndViewsTab({
 	const {objectsPortletURL} = useContext(AppContext);
 	const {
 		config,
+		openFormViewModal,
 		state: {app},
+		updateFormView,
 	} = useContext(EditAppContext);
 	const {
 		appWorkflowDataLayoutLinks: stepFormViews = [],
@@ -167,18 +165,6 @@ export default function DataAndViewsTab({
 		}
 	};
 
-	const updateFormView = (formView) => {
-		dispatchConfig({
-			formView,
-			type: UPDATE_FORM_VIEW,
-		});
-
-		dispatch({
-			...formView,
-			type: UPDATE_DATA_LAYOUT_ID,
-		});
-	};
-
 	const updateStepFormView = (formView, index) => {
 		dispatchConfig({
 			formView,
@@ -204,70 +190,6 @@ export default function DataAndViewsTab({
 		dispatch({
 			...tableView,
 			type: UPDATE_DATA_LIST_VIEW_ID,
-		});
-	};
-
-	const openFormViewModal = (
-		dataDefinitionId,
-		defaultLanguageId,
-		selectFormView,
-		dataLayoutId
-	) => {
-		const event = window.top?.Liferay.once(
-			'newFormViewCreated',
-			({dataDefinition, newFormView}) => {
-				successToast(
-					Liferay.Language.get('the-form-view-was-saved-successfully')
-				);
-				getFormViews(dataDefinitionId, defaultLanguageId).then(
-					(formViews) => {
-						const checkedFormViews = checkRequiredFields(
-							formViews,
-							dataDefinition
-						);
-
-						dispatchConfig({
-							listItems: {
-								fetching: false,
-								formViews: checkedFormViews,
-							},
-							type: UPDATE_LIST_ITEMS,
-						});
-
-						const currentFormView = checkedFormViews.find(
-							({id}) => id === newFormView.id
-						);
-
-						if (
-							!currentFormView.missingRequiredFields?.nativeField
-						) {
-							selectFormView({
-								...currentFormView,
-								name: getLocalizedValue(
-									defaultLanguageId,
-									newFormView.name
-								),
-							});
-						}
-						else if (newFormView.id === app.dataLayoutId) {
-							selectFormView({});
-						}
-					}
-				);
-			}
-		);
-		openModal({
-			onClose: () => event?.detach(),
-			title: dataLayoutId
-				? Liferay.Language.get('edit-form-view')
-				: Liferay.Language.get('new-form-view'),
-			url: `${Liferay.Util.PortletURL.createRenderURL(objectsPortletURL, {
-				dataDefinitionId,
-				dataLayoutId,
-				mvcRenderCommandName: '/app_builder/edit_form_view',
-				newCustomObject: true,
-				p_p_state: 'pop_up',
-			})}`,
 		});
 	};
 
@@ -384,9 +306,7 @@ export default function DataAndViewsTab({
 	}
 
 	return (
-		<DataAndViewsTabContext.Provider
-			value={{openFormViewModal, updateFormView}}
-		>
+		<>
 			{stepIndex > 0 ? (
 				<>
 					{duplicatedFields.length > 0 && (
@@ -629,6 +549,6 @@ export default function DataAndViewsTab({
 					)}
 				</>
 			)}
-		</DataAndViewsTabContext.Provider>
+		</>
 	);
 }
